fix(native_comp): add gcc homebrew formula dependency
Right now all *.eln files produced by native compilation links against a
GCC shared library with a absolute file path, rather than to the version
of the shared library included in the app bundle.

This means for now we need to ensure the homebrew gcc formula is
installed for native-compilation to work.

More details are available here:
https://github.com/jimeh/build-emacs-for-macos/issues/53