feat(emacs-app): update to point at new 29.1 build
The new build fixes an issue with bundled shared libraries, where
libwebp was not correctly bundled into the application. This lead to
Emacs crashing on start unless libwebp was already installed on the
target machine.

The rebuild does however increase the minimum required macOS version
from 11.0 Big Sur, to 12.0 Monterey. This is because Homebrew no longer
supports macOS 11, making it difficult to produce Emacs builds on it.

Additionally, native Apple Silicon builds are now available as well for
stable releases. They require macOS 13 Ventura however, as that is the
oldest version available on M1-based GitHub Actions runners.