revert(cask): update emacs-app-pretest to version 30.0.91-pretest
This reverts commit eb7deedd627c00a698dc809e2120148b6c6632d1.

The bump to 30.0.91-pretest happened before the intel build was ready,
meaning it was skipped in the cask file. Follow up runs haven't picked
it up cause the version of the cask was already the latest.

Hence we are reverting this, and re-running the update action for it to
correctly pick up the assets from the 30.0.91-pretest release.